backend/requirements.txt
Django==3.2.13
django-oauth-toolkit==2.0.0
djangorestframework==3.13.1
drf-yasg==1.20.0
django-cors-headers==3.11.0
djangorestframework-simplejwt